generateLabelFileData関数をutility.tsに移動
